transfere classes auxiliares de evento para arquivo event_manager.py
