+++ spf/core/models.py
@@ -1,27 +1,11 @@
-from enum import Enum
 from django.db import models
 from django.contrib.auth.models import User
 
 from opac_schema.v1.models import *
 
 
-class EventName(Enum):
-    # nomes de eventos associados à busca e ao envio de pacotes
-    RETRIEVE_PACKAGE_DATA = "RETRIVE_PACKAGE_DATA"
-    UPLOAD_NEW_PACKAGE = "UPLOAD_NEW_PACKAGE"
-    UPLOAD_EXISTING_PACKAGE = "UPLOAD_EXISTING_PACKAGE"
-
-    # nomes de eventos associados a validação
-    START_VALIDATION = "START_VALIDATION"
-    FINALIZE_VALIDATION = "FINALIZE_VALIDATION"
 
 
-class EventStatus(Enum):
-    # estado de um evento
-    INITIATED = "INITIATED"
-    PENDING = "PENDING"
-    COMPLETED = "COMPLETED"
-    FAILED = "FAILED"
 
 
 class Event(models.Model):
@@ -35,7 +19,6 @@ class Event(models.Model):
     name = models.CharField(max_length=200)
 
     # estado do evento (iniciado, realizando, concluído)
-    status = models.CharField(max_length=100, default=EventStatus.INITIATED)
 
     # informação relacionada ao evento
     annotation = models.CharField(max_length=200, null=True)
